legacy_local_ingestion/requirements.txt
requests==2.31.0
pydantic==1.10.13
python-dotenv==1.0.1